+++ src/test/Spec.hs
@@ -10,6 +10,7 @@ import Tests.Encoding (encodingJSONTests)
 import Tests.Foundry (foundryTests)
 import Tests.FoundryTestGen (foundryTestGenTests)
 import Tests.Integration (integrationTests)
+import Tests.Mutator (mutatorTests)
 import Tests.Optimization (optimizationTests)
 import Tests.Overflow (overflowTests)
 import Tests.Research (researchTests)
@@ -24,6 +25,7 @@ main = withCurrentDirectory "./tests/solidity" . defaultMain $
            [ configTests
            , compilationTests
            , seedTests
+           , mutatorTests
            , integrationTests
            , valuesTests
            , coverageTests